requirements-dev.txt
# Requirements needed for SeleniumLibrary development but not by end users.
# Creating releases has its own requirements listed in requirements-build.txt.
invoke >= 2.2.0 
mockito >= 1.4.0 
robotstatuschecker >= 3.0.1 
approvaltests == 8.4.1
empty_files >= 0.0.9
pytest >= 7.4.0 
pytest-mockito >= 0.0.4
pytest-approvaltests >= 0.2.4
requests >= 2.31.0
robotframework-pabot >= 2.16.0 
black >= 20.8b1
flake8 >= 6.1.0 

# Requirements needed when generating releases. See BUILD.rst for details.
rellu >= 0.7
twine >= 4.0.2 
wheel >= 0.41.1 
docutils >= 0.20.1 
pygments
beautifulsoup4 >= 4.12.2 

# Include normal dependencies from requirements.txt. Makes it possible to use
# requirements-dev.txt as a single requirement file in PyCharm and other IDEs.
-r requirements.txt
